Remove onDragCallback
migration from onDragCallback to useDrag finished.
the goal of the migration is to make code more react-friendly style.
previous version was based on rxJS and the lifecycle of callback was not clear.
This version use hook's state lifecycle intead which is more react-friendly style and encourage to code more reactive style.